+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0]
+
+- [web-client](./web-client) FEATURE: added the initial implementation of the components on the board
+- [web-client](./web-client) FEATURE: added mocks for the board and cards elements, matching the Frontend Mentor Figma desing
+
 ## [0.4.0]
 
 - [web-client](./web-client) FEATURE: added components related to Sidebar and bits of Header (Logo)